+++ history.md
@@ -2,5 +2,6 @@
 
   * [服务端](history/server.md)
   * [客户端](history/client.md)
+  * [嵌入式客户端](history/client-gtk.md)
   * [网页版](history/html5.md)
   * [相关组件](history/extra.md)